tests/test-color-w3c.html

<!-- 
Copyright 2013 Google Inc. All Rights Reserved.

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.

Written by Steph McArthur
-->

<!DOCTYPE html>
<style>
.anim {
  top: 50px;
  width: 100px;
  height: 100px;
  background-color: lightsteelblue;
  position: absolute;
}

#a {
  left: 50px
}

#b {
  left: 150px;
}

#c {
  left: 250px;
  background-color: green;
}

#container {
  position: relative;
  left: 0px;
  top: 50px;
  width: 400px;
  height: 200px;
  background-color: green;
}

</style>

All three boxes should end up the same color as the outline.

<div id="container">
  <div id="a" class="anim"></div>
  <div id="b" class="anim"></div>
  <div id="c" class="anim"></div>
</div>

<script src="../../web-animations-js/web-animation.js"></script>
<script src="../testharness/testharness.js"></script>
<script src='../testharness/testharnessreport.js'></script>
<script src='../extra-asserts.js'></script>
<link rel='stylesheet' href='../testharness/testharness.css'>

<script>
var colourCheck = async_test("Checking it ends up as green");
new Animation(document.querySelector("#a"), {"background-color": ["red", "green"]}, 2);
new Animation(document.querySelector("#b"), {"background-color": "rgb(0, 128, 0)"}, 2);
new Animation(document.querySelector("#c"), {"background-color": [{offset: 0.25, value: "rgb(255, 0, 0)"}, {offset: 0.75, value: "rgb(0, 0, 255)"}], operation: "add"}, 2);

setTimeout(function(){
  colourCheck.step(function(){
    var elems = document.querySelectorAll(".anim");
    for (var i = 0; i < elems.length; i++) {
      assert_color(elems[i], "green", "I'm not green!");
    }
  });
  colourCheck.done();
}, 2500);

</script>

<div id="log"></div>
